Added a BUNCH of cool plugins!
Discovered that lawrencium wreaks havoc on things, especially
with the MZScheme "big" that causes console output all over -
makes it REALLY bad! I'm still getting some of that, and I have
noticed that easytags also runs pretty regularly and may hold
things up a bit. All-in-all, though, Vim performs quite well
even with this number of plugins loaded; especially compared
with how bad it was when lawrencium was still loaded!